feat(lib): change to commonjs export
