+++ lib/__fixtures__/test.jpg
