+++ lib/__fixtures__/test.png
