+++ lib/index.js
@@ -1,9 +1,39 @@
-var gte = require("semver").gte;
+"use strict";
+const fs = require("fs");
+const promisify = require("util.promisify");
 
-var version = process.version;
+const readFile = promisify(fs.readFile);
+const writeFile = promisify(fs.writeFile);
 
-if (gte(version, "8.0.0")) {
-  module.exports = require("./index__8.0.0__.js");
-} else {
-  module.exports = require("./index__6.0.0__.js");
-}
+const plugins = [
+  [
+    "imagemin-gifsicle",
+    {
+      interlaced: true
+    }
+  ],
+  [
+    "imagemin-jpegtran",
+    {
+      progressive: true
+    }
+  ],
+  [
+    "imagemin-optipng",
+    {
+      optimizationLevel: 5
+    }
+  ],
+  [
+    "imagemin-svgo",
+    {
+      plugins: [{ removeViewBox: false }]
+    }
+  ]
+].map(([name, opts]) => require(name)(opts));
+
+const minifyFile = (exports.minifyFile = filename =>
+  [...plugins, it => writeFile(filename, it)].reduce(
+    (acc, it) => acc.then(it),
+    readFile(filename)
+  ));
